Create Primitive_Data_Types_in_Solidity_Challenge
The EtherValueConverter contract allows users to send Ether to the contract. It tracks the total amount of Ether received (receivedWei) and provides functions to retrieve this value in three different units:

getWeiValue(): Returns the total received Ether in wei.
getEtherValue(): Converts and returns the total received wei to ether.
getGweiValue(): Converts and returns the total received wei to gwei.
Users can interact with the contract to monitor and convert the accumulated Ether balance into these different units.